Add operator<< for ManaPool and update main output
Added a friend function declaration for the operator<< to the ManaPool class in ManaPool.h to allow for easy output of the ManaPool object using the << operator. Defined the operator<< function in ManaPool.cpp to iterate over the manaPool map and output its contents. Included the ManaPool.h header in World Tree.cpp to ensure the ManaPool class and its new operator<< function are available. Modified the main function in World Tree.cpp to use the overloaded << operator for outputting the mana values, replacing the previous individual std::cout statements for each mana type.